requirements.txt
flask
line-bot-sdk
openai
gspread
oauth2client
python-dotenv
gunicorn